♻️ Refactor: SortedTable component
