+++ client/src/features/search/sub/SearchSongs.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { faker } from '@faker-js/faker';
-import SortedTable from '../../../ui/Table/SortedTable.tsx';
+import SortedTable from '../../../ui/Table/SortedTable/SortedTable.tsx';
 
 interface Item {
   img: string;
